song upload made in one method

+++ src/main/java/com/example/soundcloudfinalprojectittalentss15/services/CommentService.java
@@ -4,6 +4,7 @@ import com.example.soundcloudfinalprojectittalentss15.model.DTOs.commentDTOs.Com
 import com.example.soundcloudfinalprojectittalentss15.model.entities.Comment;
 import com.example.soundcloudfinalprojectittalentss15.model.entities.Track;
 import com.example.soundcloudfinalprojectittalentss15.model.entities.User;
+import com.example.soundcloudfinalprojectittalentss15.model.exceptions.BadRequestException;
 import com.example.soundcloudfinalprojectittalentss15.model.exceptions.UnauthorizedException;
 import org.springframework.stereotype.Service;
 
@@ -64,12 +65,15 @@ public class CommentService extends AbstractService{
     public CommentInfoDTO deleteComment(int trackId, int commentId, int userId) {
         Track track = getTrackById(trackId);
         Comment comment = getCommentById(commentId);
-        User user = getUserById(userId);
+        if(trackId != comment.getTrack().getId()) {
+            throw new BadRequestException("Comment not found");
+        }
         if((track.getOwner().getId() != userId) && (comment.getUser().getId() != userId)) {
             throw new UnauthorizedException("User cannot access this operation! ");
         }
+
         CommentInfoDTO commentInfoDTO = mapper.map(comment, CommentInfoDTO.class);
         commentRepository.delete(comment);
-        return commentInfoDTO; 
+        return commentInfoDTO;
     }
 }
